requirements.txt
python-pkcs11==0.7.0
cryptography==41.0.7
pyyaml==6.0.1
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
prometheus-client==0.19.0
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0